Refactor form approach
using an iframe (~htmz) is elegant up to a point. Sadly it breaks down as soon
as we want to inspect the response or handle errors. iframe onload/onerror are
not fired for all error events (e.g. bind error / server shutdown leads to
errors that cannot be detected or handled client side. inspecting response
headers is also not possible, ...)

this approach (~triptych) is a tiny bit more manual work but handles those edge
cases.