git: Allow git push without local checkout
git is text. a lot of my backup needs are text based and i already had nice
experiences with git scraping which is similar enough. the idea here is to
process e.g. sqlite db contents in memory into text files and only push
when/what is necessary. By piggybacking on `filter=blob:none` we can skip the
local checkout if we're ok with working at the file rather than line diff lvl.